Moved resources to separate file, fixed type hints, other minor refactoring
