+++ queries/dialogues/fruitseller.toml
@@ -1,5 +1,3 @@
-dialogue_name = "fruitseller"
-
 [[resources]]
 name = "Fruits"
 type = "ListResource"
